+++ src/main/java/racing/RacingSolution.java
@@ -1,8 +1,8 @@
 package racing;
 
 import racing.controller.RacingController;
-import racing.domain.game.dto.GameRequest;
-import racing.domain.game.vo.Turns;
+import racing.domain.dto.GameRequest;
+import racing.domain.turn.Turns;
 import racing.exception.DuplicateKeyException;
 import racing.exception.EmptyCarException;
 import racing.exception.InvalidInputException;
