+++ src/main/java/racing/domain/car/Car.java
@@ -1,0 +1,8 @@
+package racing.domain.car;
+
+import racing.domain.fuel.Fuel;
+import racing.domain.Location;
+
+public interface Car {
+    Location move(Fuel fuel);
+}
