--- src/main/java/racing/domain/car/vo/fuel/Fuel.java
@@ -1,12 +1,0 @@
-package racing.domain.car.vo.fuel;
-
-@FunctionalInterface
-public interface Fuel {
-    int MIN_VALUE = 0;
-
-    int MOVABLE_VALUE = 4;
-
-    int MAX_VALUE = 9;
-
-    boolean isMovable();
-}
