+++ src/main/java/racing/domain/fuel/Fuel.java
@@ -1,0 +1,12 @@
+package racing.domain.fuel;
+
+@FunctionalInterface
+public interface Fuel {
+    int MIN_VALUE = 0;
+
+    int MOVABLE_VALUE = 4;
+
+    int MAX_VALUE = 9;
+
+    boolean isMovable();
+}
