+++ src/main/java/racing/domain/fuel/RandomFuel.java
@@ -1,0 +1,14 @@
+package racing.domain.fuel;
+
+import racing.domain.fuel.Fuel;
+
+import java.util.Random;
+
+public final class RandomFuel implements Fuel {
+    private static final Random random = new Random();
+
+    @Override
+    public boolean isMovable() {
+        return random.nextInt(Fuel.MAX_VALUE) > MOVABLE_VALUE;
+    }
+}
